[이진우] sprint7 (#99)
* initialize sprint mission5

* refactor: base code API -> RESTful API

* fix: swagger multipart_form_data, json contentType issue

* add Dockerfile

* add railway.toml

* railway.toml: dockerfile -> Dockerfile

* delete Dockerfile, railway.toml

* add Railpack.toml

* Railway.toml add env

* refactor: railway.toml

* delete build.gradle toolchain and railway.toml

* build.gradle add archiveFileName

* setting postgresql, jpa

* basic requirements

* advenced requirements

* initialize to base code

* feat: 프로필 기반 설정 및 로깅 추가

* feat: Custom Exception 추가

* feat: Validation 추가

* feat: actuator 설정

* test: UserService테스트 코드 추가

* test: ChannlService 테스트 코드 추가

* test: MessageService 테스트 코드 추가

* test: UserRepository 테스트트 코드 작성

* test: ChannelRepository 테스í트 코드 작성

* test: MessageRepositoryí 테스트 코드 추가

* fix: 중복된 config수정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -4,52 +4,50 @@ import com.sprint.mission.discodeit.dto.data.UserDto;
 import com.sprint.mission.discodeit.dto.request.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.dto.request.UserCreateRequest;
 import com.sprint.mission.discodeit.dto.request.UserUpdateRequest;
-import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
-import com.sprint.mission.discodeit.mapper.PageResponseMapper;
+import com.sprint.mission.discodeit.exception.ErrorCode;
+import com.sprint.mission.discodeit.exception.user.UserAlreadyExistsException;
+import com.sprint.mission.discodeit.exception.user.UserNotFoundException;
 import com.sprint.mission.discodeit.mapper.UserMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import com.sprint.mission.discodeit.repository.UserStatusRepository;
 import com.sprint.mission.discodeit.service.UserService;
 import com.sprint.mission.discodeit.storage.BinaryContentStorage;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 @RequiredArgsConstructor
 @Service
-@Transactional(readOnly = true)
 public class BasicUserService implements UserService {
 
     private final UserRepository userRepository;
-    private final BinaryContentRepository binaryContentRepository;
-    private final UserStatusRepository userStatusRepository;
     private final UserMapper userMapper;
-    private final BinaryContentStorage  binaryContentStorage;
-    private final PageResponseMapper<UserDto>  pageResponseMapper;
+    private final BinaryContentRepository binaryContentRepository;
+    private final BinaryContentStorage binaryContentStorage;
 
     @Transactional
     @Override
     public UserDto create(UserCreateRequest userCreateRequest,
-                       Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
+                          Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
         String username = userCreateRequest.username();
         String email = userCreateRequest.email();
 
         if (userRepository.existsByEmail(email)) {
-            throw new IllegalArgumentException("User with email " + email + " already exists");
+            throw new UserAlreadyExistsException(ErrorCode.USER_EMAIL_DUPLICATE)
+                    .addDetail("email",  email);
         }
         if (userRepository.existsByUsername(username)) {
-            throw new IllegalArgumentException("User with username " + username + " already exists");
+            throw new UserAlreadyExistsException(ErrorCode.USERNAME_DUPLICATE)
+                    .addDetail("username",  username);
         }
 
         BinaryContent nullableProfile = optionalProfileCreateRequest
@@ -59,85 +57,83 @@ public class BasicUserService implements UserService {
                     byte[] bytes = profileRequest.bytes();
                     BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
                             contentType);
-                    BinaryContent createdContent = binaryContentRepository.save(binaryContent);
-                    binaryContentStorage.put(createdContent.getId(), bytes);
-                    return createdContent;
+                    binaryContentRepository.save(binaryContent);
+                    binaryContentStorage.put(binaryContent.getId(), bytes);
+                    return binaryContent;
                 })
                 .orElse(null);
         String password = userCreateRequest.password();
 
         User user = new User(username, email, password, nullableProfile);
-
         Instant now = Instant.now();
         UserStatus userStatus = new UserStatus(user, now);
 
-        user.setStatus(userStatus);
-        User createdUser = userRepository.save(user);
-        userStatusRepository.save(userStatus);
-
-        return userMapper.toDto(createdUser);
+        User savedUser = userRepository.save(user);
+        return userMapper.toDto(savedUser);
     }
 
     @Override
     public UserDto find(UUID userId) {
         return userRepository.findById(userId)
                 .map(userMapper::toDto)
-                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+                .orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND)
+                        .addDetail("userId", userId));
     }
 
     @Override
     public List<UserDto> findAll() {
-        return userRepository.findAll().stream().map(userMapper::toDto).toList();
+        return userRepository.findAllWithProfileAndStatus()
+                .stream()
+                .map(userMapper::toDto)
+                .toList();
     }
 
     @Transactional
     @Override
     public UserDto update(UUID userId, UserUpdateRequest userUpdateRequest,
-                       Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
+                          Optional<BinaryContentCreateRequest> optionalProfileCreateRequest) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+                .orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND)
+                        .addDetail("userId", userId));
 
         String newUsername = userUpdateRequest.newUsername();
         String newEmail = userUpdateRequest.newEmail();
         if (userRepository.existsByEmail(newEmail)) {
-            throw new IllegalArgumentException("User with email " + newEmail + " already exists");
+            throw new UserAlreadyExistsException(ErrorCode.USER_EMAIL_DUPLICATE)
+                    .addDetail("email",  newEmail);
         }
         if (userRepository.existsByUsername(newUsername)) {
-            throw new IllegalArgumentException("User with username " + newUsername + " already exists");
+            throw new UserAlreadyExistsException(ErrorCode.USERNAME_DUPLICATE)
+                    .addDetail("username",  newUsername);
         }
 
         BinaryContent nullableProfile = optionalProfileCreateRequest
                 .map(profileRequest -> {
-                    Optional.ofNullable(user.getProfile())
-                            .ifPresent(profile -> binaryContentRepository.deleteById(profile.getId()));
 
                     String fileName = profileRequest.fileName();
                     String contentType = profileRequest.contentType();
                     byte[] bytes = profileRequest.bytes();
                     BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length,
                             contentType);
-                    BinaryContent createdContent = binaryContentRepository.save(binaryContent);
-                    binaryContentStorage.put(createdContent.getId(), bytes);
-
-                    return createdContent;
+                    binaryContentRepository.save(binaryContent);
+                    binaryContentStorage.put(binaryContent.getId(), bytes);
+                    return binaryContent;
                 })
                 .orElse(null);
 
         String newPassword = userUpdateRequest.newPassword();
         user.update(newUsername, newEmail, newPassword, nullableProfile);
 
-        return userMapper.toDto(userRepository.save(user));
+        return userMapper.toDto(user);
     }
 
     @Transactional
     @Override
     public void delete(UUID userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
-
-        Optional.ofNullable(user.getProfile())
-                .ifPresent(profile -> binaryContentRepository.deleteById(profile.getId()));
-        userStatusRepository.deleteByUserId(userId);
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException(ErrorCode.USER_NOT_FOUND)
+                    .addDetail("userId", userId);
+        }
 
         userRepository.deleteById(userId);
     }
